Address remaining low-risk C++ lints

+++ engine/src/modules/core/sequencer.cpp
@@ -27,7 +27,7 @@ void Sequencer::initialize(std::shared_ptr<AnthemModelBase> selfModel,
 
   // Write initial values to transport
   transport.setTicksPerQuarter(this->ticksPerQuarter());
-  transport.setBeatsPerMinute(this->beatsPerMinuteRaw() / 100.0);
+  transport.setBeatsPerMinute(static_cast<double>(this->beatsPerMinuteRaw()) / 100.0);
   transport.setActiveSequenceId(this->activeTransportSequenceID());
   transport.setActiveTrackId(this->activeTrackID());
   transport.setIsPlaying(this->isPlaying());
